backend/requirements.txt
django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
celery==5.3.1
django-celery-beat==2.5.0
redis==5.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
